refactor(orders): move work-target selection cache into logic (Refs #2277)
Extract PerPlayerWorkTargetSelectionCache and WorkTargetSelectionSnapshot to
colonizethis_logic so Flutter shell and future worker isolates share the same
pure-Dart population path. App keeps a thin re-export for existing import sites.

Update order-suggestions and UI specs for caller-owned instances. Migrate unit
tests from app to logic package.
