+++ SPEC/program/order-suggestions.md
@@ -87,7 +87,7 @@ For every suggested order `o`, appending it to the current list and validating v
 - Sim-game default AI (see [sim-game-default-ai.md](sim-game-default-ai.md)) — sim controller passes the same maps used for order validation.
 - Full AI (see [ai-systems-impl.md](ai-systems-impl.md)) — domain planners pass `tileMapByRegion` through to `suggestWorkOrders` when provided.
 - **Flutter app (`colonizethis_app`):** Riverpod providers (e.g. `availableWorkTargetIdsForUnitProvider`, optional `devExclusiveReservedWorkTileKeysProvider`) **delegate** to `colonizethis_logic` only (`getAvailableWorkTargetsForUnit` per known `unitId`). They **must not** reimplement Builder/Engineer/Merchant per-tile exclusivity or reservation rules. Reservations combine in-map `currentWork` and pending dev-exclusive work orders per [orders.md](orders.md) § WorkOrder per-tile exclusivity. **Do not** call broad `suggestWorkOrders` from panel-open or per-row Assign availability hot paths (Refs #2133).
-- **App-owned selection cache contract (`explore`, `steal_tech`, `counter_spy`, `purchase_land`, `prospect`, `build_improvement`, `upgrade_town`, `build_road`, `build_port`, `build_fort`, `build_rail`):** The app may keep a per-player work-target selection cache keyed by `(playerId, workTarget)` and exact `tileKey` members for UI target presentation. Ownership is app-only; `colonizethis_logic` remains source-of-truth for rules and validation and must not persist per-player caches across calls.
+- **Per-player work-target selection cache (`explore`, `steal_tech`, `counter_spy`, `purchase_land`, `prospect`, `build_improvement`, `upgrade_town`, `build_road`, `build_port`, `build_fort`, `build_rail`):** Tile-set materialization uses **`PerPlayerWorkTargetSelectionCache`** / **`WorkTargetSelectionSnapshot`** in **`colonizethis_logic`** (`src/orders/per_player_work_target_selection_cache.dart`) so the Flutter shell and **worker isolates** share identical population rules (Refs **#2277**). **Lifecycle:** each caller **constructs** its own cache instance and calls **`refresh`** at load / turn-resolution boundaries; logic exposes **no** global singleton for this map. Rules and validation remain in **`getValidWorkOrderTileKeysWithVisibility`** and the order engine.
 - **Cache-first selection (app shell):** For `explore`, `steal_tech`, `counter_spy`, `purchase_land`, `prospect`, `build_improvement`, `upgrade_town`, `build_road`, `build_port`, `build_fort`, and `build_rail`, the app shell reads the cached tile set when entering work-target selection mode (no live recomputation fallback in that interaction). Cache population eligibility is per-target and per-unit: `explore`, `steal_tech`, `counter_spy`, and `purchase_land` union `getValidWorkOrderTileKeysWithVisibility` across all human civilian units that support the target (same merged pattern as `explore`). Other cache-first targets include tiles from units that support the target and are `idle` with no `currentWork` and no pending work order for that unit in current draft orders.
 - **Runtime stale-tile filter for cache-first protected targets (app shell):** Before rendering cached selections for `explore`, `steal_tech`, `counter_spy`, `purchase_land`, `prospect`, `build_improvement`, `upgrade_town`, `build_road`, `build_port`, `build_fort`, and `build_rail`, the app shell applies a post-cache conflict filter (set subtraction on cached tile keys) using current draft orders and in-progress work so stale conflicting cached tiles are not selectable; this filter does not trigger live recomputation.
 
@@ -139,7 +139,7 @@ Before iterating candidate tiles, apply work-target-specific filters to dramatic
 
 | Work target | Province scope | Tile requirements |
 |-------------|----------------|-------------------|
-| `explore` | Per-player partially revealed province semantics (app may use app-owned cache) | Any tile in the province (province-level work) |
+| `explore` | Per-player partially revealed province semantics (shell may use `PerPlayerWorkTargetSelectionCache`) | Any tile in the province (province-level work) |
 | `prospect` | Land provinces (prefixed province id) | Tile must be mineral-eligible per `isMineralEligibleTile` (prospectable terrain when resolvable from tile maps; known non-mineral resources excluded); tile must **not** already appear in `WorldState.playerProspectedTiles[playerId]` |
 | `build_improvement` | Owned or purchased tiles | Tile must have a resource (`resourceByTileKey` non-empty); tile controlled by player |
 | `upgrade_town` | Owned provinces only | Province's town tile only |
