+++ SPEC/ui/civilian-units-panel.md
@@ -51,7 +51,7 @@ For each civilian unit, the panel shows:
 - **Shared row-action layout convention:** Civilian rows use the shared unit-panel row-action widget from `app/lib/features/game/widgets/units/shared/` (same abstraction used by Military/Navy rows): details stay left, row actions stay right, actions are ordered left-to-right, action group top-aligns to the first detail line, and narrow widths switch action buttons to icon-only.
 - **Dedicated locate control:** Each row includes a compact **Locate** icon button matching the naval units panel fleet locate chrome (`Icons.my_location`, `iconSize` 18, `VisualDensity.compact`; tooltip uses the shared **Locate** label). Tapping it emits **`LocateMapTileEvent` only** — the panel stays open (no **`ClosePanelEvent`**), same contract as naval’s per-fleet locate icon.
 - **Clickable row:** In **full-list** mode, tapping the row body (outside row actions) closes the panel then locates on the map (post-frame **`LocateMapTileEvent`**). In **tile-scoped** mode, tapping the row body only changes **selection**; use the locate icon to highlight/pan on the map without closing the panel.
-- **Assign (idle only):** For each unit with `Unit.status == idle` and no pending work order for this turn, the row shows an **Assign** button. Clicking it opens a menu of **allowed work orders** for that unit type only (per [civilian-units.md](../game/civilian-units.md) Work Order Summary and `workOrderTargetsByUnitType`). **Draft xor with Move:** The human shell must not keep a pending civilian **`MoveOrder`** and a **`WorkOrder`** for the **same** `unitId` in one turn; committing work via Assign clears a conflicting move in the draft helper `GameMapAreaStateLogic.addHumanWorkOrder`, matching [orders.md](../program/orders.md) § Civilian `WorkOrder` bundling (implicit move leg) and xor draft. After the user selects an order from the menu, the shell enters **work-target selection mode**: for `explore`, `steal_tech`, `counter_spy`, `purchase_land`, `prospect`, `build_improvement`, `upgrade_town`, `build_road`, `build_port`, `build_fort`, and `build_rail`, the shell reads valid target tiles from the app-owned per-player work-target selection cache (shared with province shortcut states); for other targets it computes valid tiles once using `getValidWorkOrderTileKeysWithVisibility`, passing **per-region tile maps** when loaded. For cache-first protected targets (`explore`, `steal_tech`, `counter_spy`, `purchase_land`, `prospect`, `build_improvement`, `upgrade_town`, `build_road`, `build_port`, `build_fort`, `build_rail`), the shell applies runtime stale-tile filtering on cached keys using current draft/in-progress conflicts and does not trigger live fallback recomputation in the interaction. The shell passes a **global** tile set (all regions) unchanged to the map widget. The map shows **valid target tiles** with a **flashing yellow selector** and an **orange hover cursor** (see [map-widget.md](map-widget.md) § Work target selection mode). The user may **switch region tabs** to select a target in the other region; tab changes update rendering from the same global cache and do not recompute or overwrite the cached set. Clicking a **valid** tile commits the work order and exits selection mode. For **province-level** orders (`explore`, `steal_tech`, `counter_spy`), the clicked tile is translated to the province (e.g. `regionId|provinceId|0|0`) per [orders.md](../program/orders.md). **Back-out:** Invalid/empty tile clicks are no-op (selection mode persists); explicit cancel via the top-centered prompt `cancel`, `Esc` while selection mode is active, or any left-rail map icon exits selection mode without submitting while preserving each icon's existing action.
+- **Assign (idle only):** For each unit with `Unit.status == idle` and no pending work order for this turn, the row shows an **Assign** button. Clicking it opens a menu of **allowed work orders** for that unit type only (per [civilian-units.md](../game/civilian-units.md) Work Order Summary and `workOrderTargetsByUnitType`). **Draft xor with Move:** The human shell must not keep a pending civilian **`MoveOrder`** and a **`WorkOrder`** for the **same** `unitId` in one turn; committing work via Assign clears a conflicting move in the draft helper `GameMapAreaStateLogic.addHumanWorkOrder`, matching [orders.md](../program/orders.md) § Civilian `WorkOrder` bundling (implicit move leg) and xor draft. After the user selects an order from the menu, the shell enters **work-target selection mode**: for `explore`, `steal_tech`, `counter_spy`, `purchase_land`, `prospect`, `build_improvement`, `upgrade_town`, `build_road`, `build_port`, `build_fort`, and `build_rail`, the shell reads valid target tiles from its **`PerPlayerWorkTargetSelectionCache`** instance ([order-suggestions.md](../program/order-suggestions.md) § Per-player work-target selection cache; shared semantics with province shortcut states); for other targets it computes valid tiles once using `getValidWorkOrderTileKeysWithVisibility`, passing **per-region tile maps** when loaded. For cache-first protected targets (`explore`, `steal_tech`, `counter_spy`, `purchase_land`, `prospect`, `build_improvement`, `upgrade_town`, `build_road`, `build_port`, `build_fort`, `build_rail`), the shell applies runtime stale-tile filtering on cached keys using current draft/in-progress conflicts and does not trigger live fallback recomputation in the interaction. The shell passes a **global** tile set (all regions) unchanged to the map widget. The map shows **valid target tiles** with a **flashing yellow selector** and an **orange hover cursor** (see [map-widget.md](map-widget.md) § Work target selection mode). The user may **switch region tabs** to select a target in the other region; tab changes update rendering from the same global cache and do not recompute or overwrite the cached set. Clicking a **valid** tile commits the work order and exits selection mode. For **province-level** orders (`explore`, `steal_tech`, `counter_spy`), the clicked tile is translated to the province (e.g. `regionId|provinceId|0|0`) per [orders.md](../program/orders.md). **Back-out:** Invalid/empty tile clicks are no-op (selection mode persists); explicit cancel via the top-centered prompt `cancel`, `Esc` while selection mode is active, or any left-rail map icon exits selection mode without submitting while preserving each icon's existing action.
 - **Assign (explorer shortcut mode):** In explorer-filtered shortcut mode opened from province Tile actions, pressing **Assign** on an eligible Explorer bypasses the generic choose-order menu and directly commits a pending work order targeting the already selected province-panel tile. The UI must not enter work-target selection mode for this shortcut. If assignment is no longer valid at click-time (state drift), the UI performs a silent no-op and does not commit a pending work order.
 - **Assign (builder shortcut mode):** In builder-filtered shortcut mode opened from province Tile `Build improvement`, pressing **Assign** on an eligible idle/no-pending Builder bypasses the generic choose-order menu and directly commits pending `WorkOrder(target: build_improvement, targetTileKey: <exact selected tile key>)`. Builder rows with pending/in-progress work keep standard row content and do not show Assign. If assignment is no longer valid at click-time (state drift), the UI performs a silent no-op and does not commit a pending work order.
 - **Explicit shortcut contract:** `OpenCivilianUnitsPanelEvent` carries explicit shortcut target fields; `prospectShortcutTargetTileKey` opens direct-assign `prospect`, `exploreShortcutTargetTileKey` opens direct-assign `explore`, `buildImprovementShortcutTargetTileKey` opens direct-assign `build_improvement`. At most one shortcut field is non-null for a given panel open request.
