+++ SPEC/ui/province-sea-zone-detail-overlay.md
@@ -77,7 +77,7 @@ Non-Material, pixel-art friendly: `CtPanel`, `CtTabStrip`, explicit text styles 
   - and the selected tile details are not obfuscated.
 - **Explore icon behavior:** If the human player has no Explorer units, keep icon visible but disabled (grayscale, no tap). On tap, open Civilian Units panel in explorer-only shortcut mode targeting `explore` for the exact selected tile key. Assign bypasses chooser and commits pending `WorkOrder(target: explore, targetTileKey: <exact selected tile key>)`.
 - **Explore target semantics:** `explore` remains province-level. The selected full tile key is the canonical province-level assignment key per [orders.md](../program/orders.md) and must not be rewritten to a synthetic anchor.
-- **Work-target selection cache policy:** Civilian cache-first targets and cache lifecycle match [order-suggestions.md](../program/order-suggestions.md) § App-owned selection cache contract, § Cache-first selection (app shell), and § Runtime stale-tile filter for cache-first protected targets. For **this overlay’s** province Tile section, **`explore`** and **`build_improvement`** inline actions consult that app-owned cache for the gates described below; **`Prospect with explorer`** uses stable world/player mineral and prospection state only (not selection-cache tile membership) for its icon gates. Overlay rebuild churn must not replace those cache-backed gates with live order-suggestion recomputation. Cache refresh is boundary-based: initialize on active game load/start and refresh on turn-resolution completion.
+- **Work-target selection cache policy:** Civilian cache-first targets and cache lifecycle match [order-suggestions.md](../program/order-suggestions.md) § Per-player work-target selection cache, § Cache-first selection (app shell), and § Runtime stale-tile filter for cache-first protected targets. For **this overlay’s** province Tile section, **`explore`** and **`build_improvement`** inline actions consult the shell’s **`PerPlayerWorkTargetSelectionCache`** (logic implementation) for the gates described below; **`Prospect with explorer`** uses stable world/player mineral and prospection state only (not selection-cache tile membership) for its icon gates. Overlay rebuild churn must not replace those cache-backed gates with live order-suggestion recomputation. Cache refresh is boundary-based: initialize on active game load/start and refresh on turn-resolution completion.
 - **Prospect icon gate:** Show only when tile is mineral-eligible and not already prospected by the human player, with the same visibility/obfuscation and province-only gating. Mineral eligibility matches **`isMineralEligibleTile`** in colonizethis_logic (known terrain resources such as wool on hills are not mineral-eligible).
 - **Prospect icon behavior:** If the human player has no Explorer units, keep icon visible but disabled; disabled state uses grayscale styling and no tap action.
 - Explorers with pending work still count as Explorer units for icon availability and panel filtering.
@@ -120,8 +120,8 @@ Non-Material, pixel-art friendly: `CtPanel`, `CtTabStrip`, explicit text styles 
 - Given tile details are obfuscated (`unknown` or `unrevealed`) or the selected context is sea-zone, when the overlay renders, then the UI layer does not show `Explore with explorer`.
 - Given user taps `Explore with explorer` and click-time state remains valid, when the Civilian Units panel explorer shortcut assign is triggered, then the UI layer commits pending `WorkOrder(target: explore, targetTileKey: <exact selected tile key>)` and does not enter generic work-target selection mode.
 - Given click-time state drift invalidates `Explore with explorer` assignment, when the user taps the icon, then the UI layer performs a silent no-op and commits no pending work order.
-- Given map scrolling/panning/rebuild churn with unchanged selected tile and unchanged turn snapshot, when the overlay re-renders, then **`explore`** and **`build_improvement`** Tile inline action states that are driven by the app-owned work-target selection cache read from that cache and do not perform live target-set recomputation for those gates.
-- Given turn resolution advances to the next turn snapshot, when the UI refreshes overlay-related caches, then the app-owned work-target selection cache is refreshed for subsequent overlay and shell decisions that depend on it (including **`explore`** and **`build_improvement`** Tile inline actions per the cache policy above).
+- Given map scrolling/panning/rebuild churn with unchanged selected tile and unchanged turn snapshot, when the overlay re-renders, then **`explore`** and **`build_improvement`** Tile inline action states that are driven by the shell’s **`PerPlayerWorkTargetSelectionCache`** read from that cache and do not perform live target-set recomputation for those gates.
+- Given turn resolution advances to the next turn snapshot, when the UI refreshes overlay-related caches, then **`PerPlayerWorkTargetSelectionCache`** is refreshed for subsequent overlay and shell decisions that depend on it (including **`explore`** and **`build_improvement`** Tile inline actions per the cache policy above).
 - Given province Tile details are visible and selected tile has a resource with current improvement level below the player extraction tech cap for that resource, when the overlay renders, then the UI layer shows inline `Build improvement` on the `Improvement` row.
 - Given selected tile has no resource or its current improvement level is at/above the player extraction tech cap, when the overlay renders, then the UI layer does not show `Build improvement`.
 - Given the selected context is sea-zone or tile details are obfuscated (`???`), when the overlay renders, then the UI layer does not show `Build improvement`.
